// cpo-server/src/types/Authorization.ts
export enum Entity {
  SITE = 'Site',
  SITE_AREA = 'SiteArea',
  COMPANY = 'Company',
  CHARGING_STATION = 'ChargingStation',
  CHARGING_STATION_TEMPLATE = 'ChargingStationTemplate',
  CONNECTOR = 'Connector',
  TENANT = 'Tenant',
  TRANSACTION = 'Transaction',
  REPORT = 'Report',
  USER = 'User',
  USER_SITE = 'UserSite',
  SITE_USER = 'SiteUser',
  USERS_SITES = 'UsersSites',
  LOGGING = 'Logging',
  PRICING = 'Pricing',
  PRICING_DEFINITION = 'PricingDefinition',
  BILLING = 'Billing',
  BILLING_PLATFORM = 'BillingPlatform',
  BILLING_ACCOUNT = 'BillingAccount',
  BILLING_TRANSFER = 'BillingTransfer',
  SETTING = 'Setting',
  ASYNC_TASK = 'AsyncTask',
  OCPI_ENDPOINT = 'OcpiEndpoint',
  OICP_ENDPOINT = 'OicpEndpoint',
  CONNECTION = 'Connection',
  ASSET = 'Asset',
  CAR_CATALOG = 'CarCatalog',
  CAR = 'Car',
  INVOICE = 'Invoice',
  TAX = 'Tax',
  REGISTRATION_TOKEN = 'RegistrationToken',
  CHARGING_PROFILE = 'ChargingProfile',
  NOTIFICATION = 'Notification',
  TAG = 'Tag',
  EMAID ='Emaid',
  PAYMENT_METHOD = 'PaymentMethod',
  SOURCE = 'Source',
  CONSUMPTION = 'Consumption',
  SMART_CHARGING = 'SmartCharging',
  STATISTIC = 'Statistic',
  CERTIFICATE = 'Certificate',
}

export enum Action {
  READ = 'Read',
  CREATE = 'Create',
  UPDATE = 'Update',
  UPDATE_BY_VISUAL_ID = 'UpdateByVisualID',
  REPLACE = 'Replace',
  REVOKE = 'Revoke',
  DELETE = 'Delete',
  LOGOUT = 'Logout',
  LOGIN = 'Login',
  LIST = 'List',
  IN_ERROR = 'InError',
  RESET = 'Reset',
  ASSIGN = 'Assign',
  UNASSIGN = 'Unassign',
  CLEAR_CACHE = 'ClearCache',
  TRIGGER_DATA_TRANSFER = 'DataTransfer',
  SYNCHRONIZE = 'Synchronize',
  GET_CONFIGURATION = 'GetConfiguration',
  CHANGE_CONFIGURATION = 'ChangeConfiguration',
  SYNCHRONIZE_CAR_CATALOGS = 'SynchronizeCarCatalogs',
  REMOTE_START_TRANSACTION = 'RemoteStartTransaction',
  REMOTE_STOP_TRANSACTION = 'RemoteStopTransaction',
  START_TRANSACTION = 'StartTransaction',
  STOP_TRANSACTION = 'StopTransaction',
  UNLOCK_CONNECTOR = 'UnlockConnector',
  AUTHORIZE = 'Authorize',
  SET_CHARGING_PROFILE = 'SetChargingProfile',
  GET_COMPOSITE_SCHEDULE = 'GetCompositeSchedule',
  CLEAR_CHARGING_PROFILE = 'ClearChargingProfile',
  GET_DIAGNOSTICS = 'GetDiagnostics',
  UPDATE_FIRMWARE = 'UpdateFirmware',
  EXPORT = 'Export',
  CHANGE_AVAILABILITY = 'ChangeAvailability',
  REFUND_TRANSACTION = 'RefundTransaction',
  LIST_BAGDES_EMSPS = 'TagsGetEmsp',
  SYNCHRONIZE_BILLING_USER = 'SynchronizeBillingUser',
  BILLING_SETUP_PAYMENT_METHOD = 'BillingSetupPaymentMethod',
  BILLING_PAYMENT_METHODS = 'BillingPaymentMethods',
  BILLING_DELETE_PAYMENT_METHOD = 'BillingDeletePaymentMethod',
  BILLING_CHARGE_INVOICE = 'BillingChargeInvoice',
  BILLING_ACTIVATE_ACCOUNT = 'BillingAccountActivate',
  BILLING_ONBOARD_ACCOUNT = 'BillingAccountOnboard',
  BILLING_FINALIZE_TRANSFER = 'BillingFinalizeTransfer',
  BILLING_SEND_TRANSFER = 'BillingSendTransfer',
  CHECK_CONNECTION = 'CheckConnection',
  CLEAR_BILLING_TEST_DATA = 'ClearBillingTestData',
  RETRIEVE_CONSUMPTION = 'RetrieveConsumption',
  READ_CONSUMPTION = 'ReadConsumption',
  CREATE_CONSUMPTION = 'CreateConsumption',
  PING = 'Ping',
  GENERATE_LOCAL_TOKEN = 'GenerateLocalToken',
  REGISTER = 'Register',
  TRIGGER_JOB = 'TriggerJob',
  DOWNLOAD = 'Download',
  IMPORT = 'Import',
  ASSIGN_USERS_TO_SITE = 'AssignUsersToSite',
  UNASSIGN_USERS_FROM_SITE = 'UnassignUsersFromSite',
  ASSIGN_SITES_TO_USER = 'AssignSitesToUser',
  UNASSIGN_SITES_FROM_USER = 'UnassignSitesFromUser',
  ASSIGN_UNASSIGN_USERS = 'AssignUnassignUsers',
  ASSIGN_UNASSIGN_USERS_TO_CONNECTOR = 'AssignUnassignUsersToConnector',
  ASSIGN_UNASSIGN_SITES = 'AssignUnassignSites',
  ASSIGN_ASSETS_TO_SITE_AREA = 'AssignAssets',
  UNASSIGN_ASSETS_FROM_SITE_AREA = 'UnassignAssets',
  READ_ASSETS_FROM_SITE_AREA = 'ReadAssets',
  ASSIGN_CHARGING_STATIONS_TO_SITE_AREA = 'AssignChargingStations',
  UNASSIGN_CHARGING_STATIONS_FROM_SITE_AREA = 'UnassignChargingStations',
  READ_CHARGING_STATIONS_FROM_SITE_AREA = 'ReadChargingStationsFromSiteArea',
  EXPORT_OCPP_PARAMS = 'ExportOCPPParams',
  GENERATE_QR = 'GenerateQrCode',
  MAINTAIN_PRICING_DEFINITIONS = 'MaintainPricingDefinitions',
  RESOLVE = 'Resolve',
  GET_STATUS_NOTIFICATION = 'GetStatusNotification',
  GET_BOOT_NOTIFICATION = 'GetBootNotification',
  RESERVE_NOW = 'ReserveNow',
  UPDATE_OCPP_PARAMS = 'UpdateOCPPParams',
  LIMIT_POWER = 'LimitPower',
  DELETE_CHARGING_PROFILE = 'DeleteChargingProfile',
  GET_OCPP_PARAMS = 'GetOCPPParams',
  UPDATE_CHARGING_PROFILE = 'UpdateChargingProfile',
  GET_CONNECTOR_QR_CODE = 'GetConnectorQRCode',
  VIEW_USER_DATA = 'ViewUserData',
  SYNCHRONIZE_REFUNDED_TRANSACTION = 'SynchronizeRefundedTransaction',
  PUSH_TRANSACTION_CDR = 'PushTransactionCDR',
  GET_ADVENIR_CONSUMPTION = 'GetAdvenirConsumption',
  GET_CHARGING_STATION_TRANSACTIONS = 'GetChargingStationTransactions',
  GET_ACTIVE_TRANSACTION = 'GetActiveTransaction',
  GET_COMPLETED_TRANSACTION = 'GetCompletedTransaction',
  GET_REFUNDABLE_TRANSACTION = 'GetRefundableTransaction',
  GET_REFUND_REPORT = 'GetRefundReport',
  EXPORT_COMPLETED_TRANSACTION = 'ExportCompletedTransaction',
  EXPORT_OCPI_CDR = 'ExportOcpiCdr',
  ASSIGN_UNASSIGN_USER_TO_CHARGING_STATION_CONNECTOR = 'AssignUnassignUserToChargingStationConnector'
}