// cpo-server/src/types/DataResult.ts
export interface DataResult<T> {
  count: number;
  result: T[];
  projectFields?: string[];
  metadata?: Record<string, AuthorizationDefinitionFieldMetadata>;
}
// ...
export interface CompanyDataResult extends DataResult<Company> {
  canCreate: boolean;
}